refactor: refactor auth domain
